+++ app/src/main/java/com/example/sharingang/ui/adapters/MessageAdapter.kt
@@ -9,6 +9,7 @@ import androidx.recyclerview.widget.RecyclerView
 import com.example.sharingang.R
 import com.example.sharingang.models.Chat
 import com.example.sharingang.ui.fragments.MessageFragment
+import com.example.sharingang.utils.DateHelper
 
 /**
  * MessageAdapter takes care of adapting a list of messages into a Recycler View.
@@ -56,7 +57,7 @@ class MessageAdapter(
     override fun onBindViewHolder(holder: ViewHolder, position: Int) {
         val message: String = chats[position].message
         holder.text.text = message
-        holder.time.text = chats[position].getMessageTime()
+        holder.time.text = DateHelper.formatMessageDate(chats[position].date)
     }
 
     override fun getItemCount(): Int {
